# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "post": {
        "operationId": "FacilityRequestController_createDonation",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bloodGroup": {
                    "type": "string",
                    "enum": [
                      "A+",
                      "A-",
                      "B+",
                      "B-",
                      "AB+",
                      "AB-",
                      "O+",
                      "O-"
                    ]
                  },
                  "pricePerPint": {
                    "type": "number",
                    "example": 1000
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "requiredDonors": {
                    "type": "number"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "voluntary",
                      "paid"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "open",
                      "closed"
                    ],
                    "example": "open"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donation Requests"
        ]
      },
      "get": {
        "operationId": "FacilityRequestController_getRequests",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of results to return per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "_id": "64b8f1e2c9a1f2b3d4e5f6a7",
                      "bloodGroup": "A+",
                      "quantity": 2,
                      "requiredDonors": 5,
                      "type": "voluntary"
                    }
                  ],
                  "meta": {
                    "currentPage": 1,
                    "perPage": 10,
                    "skip": 10,
                    "lastPage": 4,
                    "nextPage": 2,
                    "prevPage": null,
                    "from": 11,
                    "to": 20
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Donation Requests"
        ]
      }
    },
...
    "/donor/profile": {
      "get": {
        "operationId": "DonorController_getProfile",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "_id": "6abf49669ce610e06ce30e54",
                    "firstName": "Firstname",
                    "lastName": "Lastname",
                    "email": "newuser@localhost.com",
                    "phone": "00000000000",
                    "address": "123 Main Street",
                    "role": "donor",
                    "status": "active",
                    "donorDetails": {
                      "_id": "6abf49669ce610e06ce30e55",
                      "bloodGroup": "A+",
                      "donationStatus": "available"
                    }
                  },
                  "message": "Request made"
                }
              }
            }
          }
        },
        "tags": [
          "Donor"
        ]
      },
      "put": {
        "operationId": "DonorController_updateProfile",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "otherNames": {
                    "type": "string",
                    "example": "Smith"
                  },
                  "email": {
                    "type": "string",
                    "example": "newuser@localhost.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "00000000000"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Main Street"
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "example": 40.7128
                      },
                      "lng": {
                        "type": "number",
                        "example": -74.006
                      }
                    }
                  },
                  "bloodGroup": {
                    "type": "string",
                    "example": "A+"
                  },
                  "status": {
                    "type": "string",
                    "example": "available",
                    "enum": [
                      "available",
                      "unavailable"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "_id": "6abf49669ce610e06ce30e54",
                    "firstName": "Firstname",
                    "lastName": "Lastname",
                    "email": "newuser@localhost.com",
                    "phone": "00000000000",
                    "address": "123 Main Street",
                    "role": "donor",
                    "status": "active",
                    "donorDetails": {
                      "_id": "6abf49669ce610e06ce30e55",
                      "bloodGroup": "A+",
                      "donationStatus": "available"
                    }
                  },
                  "message": "Request made"
                }
              }
            }
          }
        },
        "tags": [
          "Donor"
        ]
      }
    },
    "/donor/availability": {
      "patch": {
        "operationId": "DonorController_updateAvailability",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "available",
                    "enum": [
                      "available",
                      "unavailable"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Donor",
          "Donor Availability - available/unavailable"
        ]
      }
    },
    "/donor/requests/get-closest-facilities": {
      "get": {
        "operationId": "DonorController_getFacilitiesClosestToLocation",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "query",
            "description": "Address of the location to search for facilities",
            "schema": {
              "example": "123 Main Street"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "Longitude of the location to search for facilities",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Latitude of the location to search for facilities",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of results to return per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "data": [
                      {
                        "_id": "6ac61236e71c125a07d3b1b7",
                        "organizationName": "string",
                        "email": "labane3138@meinvr.com",
                        "phone": "string",
                        "address": "string",
                        "location": {
                          "lat": 0,
                          "lng": 0
                        },
                        "role": "facility"
                      }
                    ],
                    "meta": {
                      "currentPage": "1",
                      "perPage": 10,
                      "skip": 0,
                      "lastPage": 1,
                      "nextPage": null,
                      "prevPage": null,
                      "from": 1,
                      "to": 1
                    }
                  },
                  "message": "Request made"
                }
              }
            }
          }
        },
        "tags": [
          "Donor"
        ]
      }
    },
    "/donor/requests": {
      "get": {
        "operationId": "DonorController_getRequests",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status of the requests to filter by",
            "schema": {
              "enum": [
                "open",
                "closed"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results to return per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "data": [
                      {
                        "_id": "6ac6503a53d954574f151762",
                        "bloodGroup": "A+",
                        "status": "closed",
                        "quantity": 2,
                        "pricePerPint": 0,
                        "requiredDonors": 1,
                        "type": "voluntary",
                        "facility": {
                          "_id": "6ac61236e71c125a07d3b1b7",
                          "organizationName": "string",
                          "email": "labane3138@meinvr.com",
                          "phone": "string",
                          "id": "6ac61236e71c125a07d3b1b7"
                        },
                        "createdAt": "2026-10-07T13:59:22.755Z",
                        "updatedAt": "2026-10-07T14:02:29.283Z",
                        "__v": 0
                      }
                    ],
                    "meta": {
                      "currentPage": "1",
                      "perPage": 10,
                      "skip": 0,
                      "lastPage": 1,
                      "nextPage": null,
                      "prevPage": null,
                      "from": 1,
                      "to": 1
                    }
                  },
                  "message": "Request made"
                }
              }
            }
          }
        },
        "tags": [
          "Donor"
        ]
      }
    },
    "/donor/requests/{requestId}/respond": {
      "post": {
        "operationId": "DonorController_respondToRequest",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "description": "ID of the request to respond to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donor"
        ]
      }
    },
    "/donor/donations": {
      "get": {
        "operationId": "DonorController_getDonations",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status of the requests to filter by",
            "schema": {
              "enum": [
                "pending",
                "accepted",
                "paid",
                "rejected"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results to return per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token for authentication",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Donor"
        ]
      }
    },
...
  },
  "info": {
    "title": "BloodBridge API",
    "description": "BloodBridge helps verified donors, hospitals, and blood banks respond to critical blood needs quickly, safely, and efficiently. Our direct coordination pipeline reduces delays in medical emergencies.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization"
      }
    },
...
  }
}